requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
htmldocx==0.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
numpy==1.19.5
pandas==1.2.0
pylint==2.6.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
